Thirteenth Commit

+++ backend/controllers/cartController.js
@@ -5,16 +5,17 @@ const cartController = async(req,res) => {
 
     const{id,name,price,image,quantity} = req.body
 
-    // if (!id || !name || !price || !image || !quantity) {
-    //     return res.status(401).json({message:'fill all the fields', status:401, success:false})
-    // }
+    if (!id || !name || !price || !image || !quantity) {
+        return res.status(401).json({message:'fill all the fields', status:401, success:false})
+    }
 
     const product = new CART ({
         id,
         name,
         price,
         image,
-        quantity
+        quantity:1,
+        userId:req.user._id
     })
 
     await product.save()
@@ -25,7 +26,7 @@ const cartController = async(req,res) => {
 
 const getCartProduct = async(req,res) => {
 
-    const products = await CART.find()
+    const products = await CART.find({userId:req.user._id})
 
     if (!products) {
         return res.status(404).json({message:"No Product Found", status:404, success:false})
@@ -39,7 +40,7 @@ const addUpdateCartProduct = async(req,res) => {
     const{id} = req.body
 
     const updatedCart = await CART.findOneAndUpdate(
-      { id },
+      { id, userId:req.user._id },
       { $inc: { quantity: 1 } },
       { new: true, runValidators: true }
     );
@@ -61,7 +62,7 @@ const deleteUpdateCartProduct = async(req,res) => {
     const{id} = req.body
 
     const updatedCart = await CART.findOneAndUpdate(
-      { id },
+      { id, userId:req.user._id },
       { $inc: { quantity: -1 } },
       { new: true, runValidators: true }
     );
@@ -82,7 +83,7 @@ const deleteProduct = async(req,res) => {
 
     let {id} = req.body
 
-    await CART.findOneAndDelete({id})
+    await CART.findOneAndDelete({id, userId:req.user._id})
 
     return res.status(200).json({message:'Product Deleted Successfully!!', status:200, success:true})
